requirements.txt
flask==3.0.2
openai==1.56.1
python-dotenv==1.0.1
httpx>=0.24.1
